refactor: reorganize OIDC utility functions and constants
